+++ src/ua/dp/drizhiruk/lesson2/sorting/Sort.java
@@ -1,0 +1,6 @@
+package ua.dp.drizhiruk.lesson2.sorting;
+
+public class Sort {
+    public void sort(int[] ints) {
+    }
+}
